+++ config/configyaml/run/run.go
@@ -1,7 +1,6 @@
 package run
 
 import (
-	"github.com/pkg/errors"
 	"github.com/rliebz/tusk/config/configyaml/marshal"
 	"github.com/rliebz/tusk/config/configyaml/when"
 )
@@ -16,20 +15,23 @@ type Run struct {
 // UnmarshalYAML allows plain strings to represent a run struct. The value of
 // the string is used as the Command field.
 func (r *Run) UnmarshalYAML(unmarshal func(interface{}) error) error {
+
+	var err error
+
 	var command string
-	if err := unmarshal(&command); err == nil {
+	if err = unmarshal(&command); err == nil {
 		*r = Run{Command: marshal.StringList{command}}
 		return nil
 	}
 
 	type runType Run // Use new type to avoid recursion
 	var runItem *runType
-	if err := unmarshal(&runItem); err == nil {
+	if err = unmarshal(&runItem); err == nil {
 		*r = *(*Run)(runItem)
 		return nil
 	}
 
-	return errors.New("could not parse run item")
+	return err
 }
 
 // List is a list of run items with custom yaml unmarshalling.
@@ -37,17 +39,20 @@ type List []*Run
 
 // UnmarshalYAML allows single items to be used as lists.
 func (rl *List) UnmarshalYAML(unmarshal func(interface{}) error) error {
-	var runItem *Run
-	if err := unmarshal(&runItem); err == nil {
-		*rl = List{runItem}
-		return nil
-	}
+
+	var err error
 
 	var runSlice []*Run
-	if err := unmarshal(&runSlice); err == nil {
+	if err = unmarshal(&runSlice); err == nil {
 		*rl = runSlice
 		return nil
 	}
 
-	return errors.New("could not parse runlist")
+	var runItem *Run
+	if err = unmarshal(&runItem); err == nil {
+		*rl = List{runItem}
+		return nil
+	}
+
+	return err
 }