+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 This project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Changed
+- Improve error handling in custom yaml unmarshalling.
 
 ## 0.1.1 (2017-09-16)
 ### Changed