+++ CHANGELOG.md
@@ -9,6 +9,7 @@ This project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html
 - Improve error handling in custom yaml unmarshalling.
 - Exit for unexpected errors occuring during when clause validation.
 - Use exit code 1 for all unexpected errors.
+- Unexpected arguments now cause an error.
 
 ### Fixed
 - Short names cannot exceed one character in length.